use golangci-lint to clean up some things

+++ internal/adapter/convert_twelve.go
@@ -16,9 +16,9 @@ func ConvertTwelveTask(
 	for i := 0; i < len(task); i++ {
 		b = task[i]
 		if b > '0' && b <= '9' {
-			output[r][c] = uint8(b - '0')
+			output[r][c] = b - '0'
 			if output[r][c] == 1 && i+1 < len(task) && task[i+1] >= '0' && task[i+1] <= '2' {
-				output[r][c] = 10 + uint8(task[i+1]-'0')
+				output[r][c] = 10 + task[i+1] - '0'
 				i++
 			}
 		} else if b == '_' {
